requirements_group_2.txt
openai
whisper
SpeechRecognition
nltk
python-dotenv
json